+++ src/main.cpp
@@ -727,134 +727,3 @@ void MainCleanup()
         orphanpool.mapOrphanTransactionsByPrev.clear();
     }
 }
-
-// Version is current unix epoch time. Nov 1, 2018 at 12am
-static const uint64_t MEMPOOL_DUMP_VERSION = 1541030400;
-
-bool LoadMempool(void)
-{
-    int64_t nExpiryTimeout = GetArg("-mempoolexpiry", DEFAULT_MEMPOOL_EXPIRY) * 60 * 60;
-    FILE *filestr = fopen((GetDataDir() / "mempool.dat").string().c_str(), "rb");
-    CAutoFile file(filestr, SER_DISK, CLIENT_VERSION);
-    if (file.IsNull())
-    {
-        LOGA("Failed to open mempool file from disk. Continuing anyway.\n");
-        return false;
-    }
-
-    int64_t count = 0;
-    int64_t skipped = 0;
-    int64_t nNow = GetTime();
-
-    try
-    {
-        uint64_t version;
-        file >> version;
-        if (version != MEMPOOL_DUMP_VERSION)
-        {
-            return false;
-        }
-        uint64_t num;
-        file >> num;
-        double prioritydummy = 0;
-        while (num--)
-        {
-            CTransaction tx;
-            int64_t nTime;
-            int64_t nFeeDelta;
-            file >> tx;
-            file >> nTime;
-            file >> nFeeDelta;
-
-            CAmount amountdelta = nFeeDelta;
-            if (amountdelta)
-            {
-                mempool.PrioritiseTransaction(tx.GetHash(), tx.GetHash().ToString(), prioritydummy, amountdelta);
-            }
-            if (nTime + nExpiryTimeout > nNow)
-            {
-                CTxInputData txd;
-                txd.tx = MakeTransactionRef(tx);
-                EnqueueTxForAdmission(txd);
-                ++count;
-            }
-            else
-            {
-                ++skipped;
-            }
-
-            if (ShutdownRequested())
-                return false;
-        }
-        std::map<uint256, CAmount> mapDeltas;
-        file >> mapDeltas;
-
-        for (const auto &i : mapDeltas)
-        {
-            mempool.PrioritiseTransaction(i.first, i.first.ToString(), prioritydummy, i.second);
-        }
-    }
-    catch (const std::exception &e)
-    {
-        LOGA("Failed to deserialize mempool data on disk: %s. Continuing anyway.\n", e.what());
-        return false;
-    }
-
-    LOGA("Imported mempool transactions from disk: %i successes, %i expired\n", count, skipped);
-    return true;
-}
-
-bool DumpMempool(void)
-{
-    int64_t start = GetTimeMicros();
-
-    std::map<uint256, CAmount> mapDeltas;
-    std::vector<TxMempoolInfo> vInfo;
-
-    {
-        READLOCK(mempool.cs);
-        for (const auto &i : mempool.mapDeltas)
-        {
-            mapDeltas[i.first] = i.second.first;
-        }
-        vInfo = mempool.AllTxMempoolInfo();
-    }
-
-    int64_t mid = GetTimeMicros();
-
-    try
-    {
-        FILE *filestr = fopen((GetDataDir() / "mempool.dat.new").string().c_str(), "wb");
-        if (!filestr)
-        {
-            return false;
-        }
-
-        CAutoFile file(filestr, SER_DISK, CLIENT_VERSION);
-
-        uint64_t version = MEMPOOL_DUMP_VERSION;
-        file << version;
-
-        file << (uint64_t)vInfo.size();
-        for (const auto &i : vInfo)
-        {
-            file << *(i.tx);
-            file << (int64_t)i.nTime;
-            file << (int64_t)i.feeDelta;
-            mapDeltas.erase(i.tx->GetHash());
-        }
-
-        file << mapDeltas;
-        FileCommit(file.Get());
-        file.fclose();
-        RenameOver(GetDataDir() / "mempool.dat.new", GetDataDir() / "mempool.dat");
-        int64_t last = GetTimeMicros();
-        LOGA("Dumped mempool: %gs to copy, %gs to dump\n", (mid - start) * 0.000001, (last - mid) * 0.000001);
-    }
-    catch (const std::exception &e)
-    {
-        LOGA("Failed to dump mempool: %s. Continuing anyway.\n", e.what());
-        return false;
-    }
-    return true;
-}
